+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/TablePopperMediaDialogController.java
@@ -393,7 +393,8 @@ public class TablePopperMediaDialogController implements Initializable, DialogCo
     boolean append = false;
 
     ObservableList<GameMediaItemRepresentation> items = assetList.getItems();
-    boolean alreadyExists = items.stream().anyMatch(i -> i.getName().equalsIgnoreCase(tableAsset.toAssetName(game.getGameDisplayName())));
+    String targetName = game.getGameName() + "." + FilenameUtils.getExtension(tableAsset.getName());
+    boolean alreadyExists = items.stream().anyMatch(i -> i.getName().equalsIgnoreCase(targetName));
     if (alreadyExists) {
       Optional<ButtonType> buttonType = WidgetFactory.showConfirmationWithOption(Studio.stage, "Asset Exists", "An asset with the same name already exists.",
         "Overwrite existing asset or append new asset?", "Overwrite", "Append");